Refactor: improve code readablility in RecentTodos component
